refactor(project): refactor updating form
refactor: updating form
